cmake.sh * allow in source source build(s) * also added _F_cmake_in_source_build

+++ source/include/cmake.sh
@@ -29,6 +29,10 @@ if [ -z "$_F_cmake_color" ]; then
 	_F_cmake_color="OFF"
 fi
 
+if [ -z "$_F_cmake_in_source_build" ]; then
+	_F_cmake_in_source_build=0
+fi
+
 ###
 # == APPENDED VARIABLES
 # * cmake to makedepends()
@@ -46,6 +50,12 @@ CMake_conf()
 	## LIB_INSTALL_DIR -> libdir
 	## LOCALSTATE_INSTALL_DIR -> localstatedir
 
+	if [ "$_F_cmake_in_source_build" -eq "0" ]; then
+		_F_cmake_src=".."
+	else
+		_F_cmake_src="."
+	fi
+
 	cmake \
 		-DCMAKE_INSTALL_PREFIX=/usr \
 		-DSYSCONF_INSTALL_DIR=/etc \
@@ -56,7 +66,7 @@ CMake_conf()
 		-DCMAKE_VERBOSE_MAKEFILE="$_F_cmake_verbose" \
 		-DCMAKE_COLOR_MAKEFILE="$_F_cmake_color" \
 		-DCMAKE_EXE_LINKER_FLAGS="$LDFLAGS" \
-		$_F_cmake_confopts "$@" .. || Fdie
+		$_F_cmake_confopts "$@" $_F_cmake_src || Fdie
 }
 
 ###
@@ -64,15 +74,20 @@ CMake_conf()
 ###
 CMake_prepare_build()
 {
-	Fcd
+	Fcd $_F_cd_path
+	echo $pwd
 	Fpatchall
-	if [ -d  "build" ]; then
-		rm -rf build || Fdie
-		mkdir build || Fdie
-		cd build || Fdie
+	if [ "$_F_cmake_in_source_build" -eq "0" ]; then
+		if [ -d  "build" ]; then
+			rm -rf build || Fdie
+			mkdir build || Fdie
+			cd build || Fdie
+		else
+			mkdir build || Fdie
+			cd build || Fdie
+		fi
 	else
-		mkdir build || Fdie
-		cd build || Fdie
+		export CMAKE_IN_SOURCE_BUILD=1
 	fi
 }
